Changed kvp field in Glue instance table from int to float

+++ metadata-index/backend/datalake_tables_config.py
@@ -484,7 +484,7 @@ instance_table_columns = [
   },
   {
     "Name": "kvp",
-    "Type": "int",
+    "Type": "float",
     "Comment": ""
   },
   {
